Add sys/avr/lib
